Extract prepareTotalItems to TS file
