+++ assets/js/base/context/hooks/payment-methods/use-payment-method-interface.js
@@ -19,62 +19,11 @@ import { useCheckoutContext } from '../../providers/cart-checkout/checkout-state
 import { usePaymentMethodDataContext } from '../../providers/cart-checkout/payment-methods';
 import { useShippingDataContext } from '../../providers/cart-checkout/shipping';
 import { useCustomerDataContext } from '../../providers/cart-checkout/customer';
+import { prepareTotalItems } from './utils';
 
 /**
  * @typedef {import('@woocommerce/type-defs/registered-payment-method-props').RegisteredPaymentMethodProps} RegisteredPaymentMethodProps
- * @typedef {import('@woocommerce/type-defs/cart').CartTotalItem} CartTotalItem
  */
-
-/**
- * Prepares the total items into a shape usable for display as passed on to
- * registered payment methods.
- *
- * @param {Object} totals Current cart total items
- * @param {boolean} needsShipping Whether or not shipping is needed.
- *
- * @return {CartTotalItem[]}  Array for cart total items prepared for use.
- */
-export const prepareTotalItems = ( totals, needsShipping ) => {
-	const newTotals = [];
-	const factory = ( label, property ) => {
-		const value = parseInt( totals[ property ], 10 );
-		const tax = parseInt( totals[ property + '_tax' ], 10 );
-		return {
-			label,
-			value,
-			valueWithTax: value + tax,
-		};
-	};
-	newTotals.push(
-		factory(
-			__( 'Subtotal:', 'woo-gutenberg-products-block' ),
-			'total_items'
-		)
-	);
-	newTotals.push(
-		factory( __( 'Fees:', 'woo-gutenberg-products-block' ), 'total_fees' )
-	);
-	newTotals.push(
-		factory(
-			__( 'Discount:', 'woo-gutenberg-products-block' ),
-			'total_discount'
-		)
-	);
-	newTotals.push( {
-		label: __( 'Taxes:', 'woo-gutenberg-products-block' ),
-		value: parseInt( totals.total_tax, 10 ),
-		valueWithTax: parseInt( totals.total_tax, 10 ),
-	} );
-	if ( needsShipping ) {
-		newTotals.push(
-			factory(
-				__( 'Shipping:', 'woo-gutenberg-products-block' ),
-				'total_shipping'
-			)
-		);
-	}
-	return newTotals;
-};
 
 /**
  * @return {RegisteredPaymentMethodProps} Interface to use as payment method props.
